# api/docs/swagger.yaml
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  github_com_USACE_instrumentation-api_api_internal_model.PlotConfig:
    properties:
      auto_range:
        type: boolean
      create_date:
        type: string
      creator:
        type: string
      date_range:
        type: string
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      show_comments:
        type: boolean
      show_masked:
        type: boolean
      show_nonvalidated:
        type: boolean
      slug:
        type: string
      threshold:
        type: integer
      timeseries_id:
        items:
          type: string
        type: array
      update_date:
        type: string
      updater:
        type: string
    type: object
  github_com_USACE_instrumentation-api_api_internal_model.Profile:
    properties:
      email:
        type: string
      id:
        type: string
      is_admin:
        type: boolean
      roles:
        items:
          type: string
        type: array
      tokens:
        items:
          $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.TokenInfoProfile'
        type: array
      username:
        type: string
    type: object
...
  github_com_USACE_instrumentation-api_api_internal_model.Token:
    properties:
      issued:
        type: string
      profile_id:
        type: string
      secret_token:
        type: string
      token_id:
        type: string
    type: object
  github_com_USACE_instrumentation-api_api_internal_model.TokenInfoProfile:
    properties:
      issued:
        type: string
      token_id:
        type: string
    type: object
...
info:
  contact: {}
  description: Monitoring Instrumentation Data Acquisition Systems (MIDAS) Web API
  license:
    name: MIT
    url: https://github.com/USACE/instrumentation-api/blob/555ea51191ff1245fe5910a295862be7514aaec6/LICENSE.md
  title: MIDAS Web API
  version: "2.0"
paths:
...
  /my_profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.Profile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: gets profile for current authenticated user
      tags:
      - profile
...
  /my_tokens:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.Token'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: creates token for a profile
      tags:
      - profile
  /my_tokens/{token_id}:
    delete:
      parameters:
      - description: token uuid
        format: uuid
        in: path
        name: token_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: deletes a token for a profile
      tags:
      - profile
...
  /profiles:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.Profile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: creates a user profile
      tags:
      - profile
...
  /projects/{project_id}/plot_configurations:
    get:
      parameters:
      - description: project uuid
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.PlotConfig'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: lists plot configs
      tags:
      - plot-config
    post:
      parameters:
      - description: project uuid
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: plot config payload
        in: body
        name: plot_config
        required: true
        schema:
          $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.PlotConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.PlotConfig'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      tags:
      - plot-config
  /projects/{project_id}/plot_configurations/{plot_configuration_id}:
    delete:
      parameters:
      - description: plot config uuid
        format: uuid
        in: path
        name: plot_configuration_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      tags:
      - plot-config
    get:
      parameters:
      - description: project uuid
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: plot config uuid
        format: uuid
        in: path
        name: plot_configuration_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.PlotConfig'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      tags:
      - plot-config
    put:
      parameters:
      - description: project uuid
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: plot config payload
        in: body
        name: plot_config
        required: true
        schema:
          $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.PlotConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.PlotConfig'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      tags:
      - plot-config
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"